add models from our ER draft, add readme
The initial migration was overwritten, so you should delete your local
db and let django recreate that.

The commit contains a new User entity which is used as the default for
authentication. It inherits all the requires traits from the framework
supplied User class.

The readme now contains basic development procedured including how to
create a new entity.